pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robotframework-okw-remote-ssh"
version = "0.1.0"
description = "Standalone Robot Framework library for deterministic remote interaction via SSH."
authors = [{ name = "Zoltan Hrabovszki" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "robotframework>=6.0",
  "paramiko>=3.0",
  "PyYAML>=6.0",
  "okw-contract-utils>=0.1.0"
]

[tool.setuptools.packages.find]
where = ["src"]
